+++ service/config/runtimeconfig.go
@@ -9,10 +9,9 @@ import (
 // RuntimeConfig is a global configuration for the service.
 // It is passed down and should not be modified unless cloned.
 type RuntimeConfig struct {
-	CacheSaveInterval uint64
+	StateBundleSize uint64
 
 	MaxWasmFuel                uint64 // if not 0, enable fuel consumption monitoring to stop runaway wasm module processing forever
-	SubrequestsSplitSize       uint64 // in multiple of the SaveIntervals above
 	MaxJobsAhead               uint64 // limit execution of depencency jobs so they don't go too far ahead of the modules that depend on them (ex: module X is 2 million blocks ahead of module Y that depends on it, we don't want to schedule more module X jobs until Y caught up a little bit)
 	DefaultParallelSubrequests uint64 // how many sub-jobs to launch for a given user
 	// derives substores `states/`, for `store` modules snapshots (full and partial)
@@ -26,8 +25,7 @@ type RuntimeConfig struct {
 }
 
 func NewRuntimeConfig(
-	cacheSaveInterval uint64,
-	subrequestsSplitSize uint64,
+	stateBundleSize uint64,
 	parallelSubrequests uint64,
 	maxJobsAhead uint64,
 	maxWasmFuel uint64,
@@ -36,8 +34,7 @@ func NewRuntimeConfig(
 	workerFactory work.WorkerFactory,
 ) RuntimeConfig {
 	return RuntimeConfig{
-		CacheSaveInterval:          cacheSaveInterval,
-		SubrequestsSplitSize:       subrequestsSplitSize,
+		StateBundleSize:            stateBundleSize,
 		DefaultParallelSubrequests: parallelSubrequests,
 		MaxJobsAhead:               maxJobsAhead,
 		MaxWasmFuel:                maxWasmFuel,